+++ src/AccesoaDatos/PasajeData.java
@@ -31,7 +31,7 @@ public class PasajeData {
         try {
             PreparedStatement ps = red.prepareStatement(sql, Statement.RETURN_GENERATED_KEYS);
             ps.setString(1, pasaje.getTipoTransporte());
-            ps.setBigDecimal(2, pasaje.getImporte());
+            ps.setDouble(2, pasaje.getImporte());
             ps.setInt(3, pasaje.getCiudadOrigen().getId_ciudad());
             ps.setInt(4, pasaje.getCiudadDestino().getId_ciudad());
             ps.setBoolean(5, pasaje.isEstado());
@@ -79,7 +79,7 @@ public class PasajeData {
               p.setId_pasaje(id);
               
               p.setCiudadOrigen(c);
-              p.setImporte(rs.getBigDecimal("importe"));
+              p.setImporte(rs.getDouble("importe"));
               p.setTipoTransporte(rs.getString("tipo_Transporte"));
               p.setEstado(rs.getBoolean("estado"));
               
@@ -104,12 +104,12 @@ public class PasajeData {
             Ciudad c = cd.buscarCiudadporid(rs.getInt("id_ciudadOrigen"));
             p.setId_pasaje(rs.getInt("id_pasaje"));
             p.setCiudadOrigen(c);
-            p.setImporte(rs.getBigDecimal("importe"));
+            p.setImporte(rs.getDouble("importe"));
             p.setTipoTransporte(rs.getString("tipo_Transporte"));
             p.setEstado(rs.getBoolean("estado"));
         }
     } catch (SQLException e) {
-        // Manejo de excepciones
+        
     }
     return p;
 }
@@ -135,7 +135,7 @@ public class PasajeData {
             p.setId_pasaje(rs.getInt("id_pasaje"));
             p.setCiudadOrigen(cOrigen);
             p.setCiudadDestino(cDestino);
-            p.setImporte(rs.getBigDecimal("importe"));
+            p.setImporte(rs.getDouble("importe"));
             p.setTipoTransporte(rs.getString("tipo_Transporte"));
             p.setEstado(rs.getBoolean("estado"));
             
@@ -172,7 +172,7 @@ public class PasajeData {
     try {
         PreparedStatement ps = red.prepareStatement(sql);
         ps.setString(1, pasaje.getTipoTransporte());
-        ps.setBigDecimal(2, pasaje.getImporte());
+        ps.setDouble(2, pasaje.getImporte());
         ps.setBoolean(3, pasaje.isEstado());
         ps.setInt(4, id);
